+++ db/migrate/20160906145713_insert_custom_aggregations.rb
@@ -1,5 +1,5 @@
 class InsertCustomAggregations < ActiveRecord::Migration
-  def change
+  def up
     # available reservations hours for machines
     machine = StatisticIndex.find_by_es_type_key('machine')
     machine_hours = StatisticType.find_by(key: 'hour', statistic_index_id: machine.id)
@@ -25,5 +25,18 @@ class InsertCustomAggregations < ActiveRecord::Migration
       query: '{"size":0, "aggregations":{"%{aggs_name}":{"sum":{"field":"nb_total_places"}}}, "query":{"bool":{"must":[{"range":{"start_at":{"gte":"%{start_date}", "lte":"%{end_date}"}}}, {"match":{"available_type":"training"}}]}}}'
     })
     available_tickets.save!
+  end
+
+  def down
+
+    machine = StatisticIndex.find_by_es_type_key('machine')
+    machine_hours = StatisticType.find_by(key: 'hour', statistic_index_id: machine.id)
+
+    StatisticCustomAggregation.where(field: 'available_hours', statistic_type_id: machine_hours.id).first.destroy!
+
+    training = StatisticIndex.find_by_es_type_key('training')
+    training_bookings = StatisticType.find_by(key: 'booking', statistic_index_id: training.id)
+
+    StatisticCustomAggregation.where(field: 'available_tickets', statistic_type_id: training_bookings.id).first.destroy!
   end
 end